+++ sbuildr/project/file_manager.py
@@ -88,7 +88,7 @@ class FileManager(object):
     # If the file exists but is not in this FileManager's tracked directories, returns an empty list.
     # The returned list is in order of proximity to the root. The first element is closest to the root.
     def find(self, path: str) -> List[str]:
-        # TODO: endswith is not a good way to do this.
+        # TODO: endswith is not a good way to do this. Need a path-separator-aware endswith.
         candidates = set([fpath for fpath in self.files if fpath.endswith(path)])
         # Prefer shorter paths, i.e. closer to the root.
         candidates = list(sorted(candidates, key=lambda elem: len(elem)))